fix: AppInfo 모델의 설치 수 관련 필드들 정수형 사이즈 초과로 인한 오 수정
https://trello.com/c/V6iqR608

+++ app/src/main/java/com/formakers/fomes/common/view/custom/RecommendAppItemView.java
@@ -116,7 +116,7 @@ public class RecommendAppItemView extends ConstraintLayout {
         setWishListChecked(appInfo.getWishedByMe());
     }
 
-    private void setVerboseGroup(Double star, Integer installsMin, String contentsRating) {
+    private void setVerboseGroup(Double star, Long installsMin, String contentsRating) {
         reviewScoreTextView.setText(String.format(getContext().getString(R.string.format_app_info_star_score),
                 (float) Math.round(star * 10) / 10));
         downloadCountTextView.setText(String.format(getContext().getString(R.string.format_app_info_download_count),
